Update -- material ui changes.

+++ lib/feature/activity/payment_activity_transaction/reject_request_payment_screens/reject_request_payment_page_view.dart
@@ -36,7 +36,7 @@ class RejectRequestPaymentPageView extends BasePageViewWidget<RejectRequestPayme
               dataBuilder: (context, currentPage) {
                 return DotsIndicator(
                   dotsCount: pages.length,
-                  position: currentPage!.toDouble(),
+                  position: currentPage!,
                   mainAxisSize: MainAxisSize.max,
                   decorator: DotsDecorator(
                       shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(5)),
