+++ lib/feature/activity/payment_activity_transaction/return_payment_transaction/return_payment_transaction_slider_page_view.dart
@@ -45,7 +45,7 @@ class ReturnPaymentTransactionSliderPageView
               dataBuilder: (context, currentPage) {
                 return DotsIndicator(
                   dotsCount: pages.length,
-                  position: currentPage!.toDouble(),
+                  position: currentPage!,
                   mainAxisSize: MainAxisSize.max,
                   decorator: DotsDecorator(
                       shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(5)),
