+++ lib/feature/supplementary_debit_card/supplementary_debit_card_page_view.dart
@@ -38,7 +38,7 @@ class SupplementaryDebitCardPageView extends BasePageViewWidget<SupplementaryDeb
               dataBuilder: (context, currentPage) {
                 return DotsIndicator(
                   dotsCount: pages.length,
-                  position: currentPage!.toDouble(),
+                  position: currentPage!,
                   mainAxisSize: MainAxisSize.max,
                   decorator: DotsDecorator(
                       shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(5)),
