+++ lib/feature/supplementary_debit_card_pin_set/supplementary_debit_card_pin_set_page_view.dart
@@ -35,7 +35,7 @@ class SupplementaryDebitCardPinSetPageView
               dataBuilder: (context, currentPage) {
                 return DotsIndicator(
                   dotsCount: pages.length,
-                  position: currentPage!.toDouble(),
+                  position: currentPage!,
                   mainAxisSize: MainAxisSize.max,
                   decorator: DotsDecorator(
                       shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(5)),
